+++ src/main/java/edu/suffolk/litlab/efspserver/services/acme/AcmeChallengePublisher.java
@@ -1,4 +1,4 @@
-package edu.suffolk.litlab.efspserver.services;
+package edu.suffolk.litlab.efspserver.services.acme;
 
 import org.shredzone.acme4j.challenge.Http01Challenge;
