+++ src/main/java/edu/suffolk/litlab/efspserver/services/acme/AcmeChallengeWriter.java
@@ -1,4 +1,4 @@
-package edu.suffolk.litlab.efspserver.services;
+package edu.suffolk.litlab.efspserver.services.acme;
 
 import java.io.File;
 import java.io.FileWriter;
